requirements.txt

Django
asgiref
sqlparse
pytz
psycopg2-binary
dj-database-url
djangorestframework
djangorestframework_simplejwt
drf-spectacular
django-cors-headers
whitenoise
django-db-connection-pool
django-redis
redis
PyJWT
python-decouple
python-dotenv
pillow
stripe
django-extensions